DEMO Simulation/requirements.txt
# Core numerical and scientific computing
numpy>=1.21.0
scipy>=1.7.0
pandas>=1.3.0

# Visualization
matplotlib>=3.4.0
networkx>=2.6.0
plotly>=5.3.0
seaborn>=0.11.0

# Configuration and environment
python-dotenv>=0.19.0
pyyaml>=5.4.0

# Progress bars and logging
tqdm>=4.62.0
colorama>=0.4.4
loguru>=0.5.3

# Testing
pytest>=6.2.0
pytest-cov>=2.12.0
pytest-xdist>=2.4.0

# Performance monitoring
psutil>=5.8.0
memory-profiler>=0.58.0

# Data serialization
json5>=0.9.6
orjson>=3.6.0

# Async support (for future extensions)
aiohttp>=3.8.0
asyncio>=3.4.3

# Web dashboard (optional, for web interface)
flask>=2.0.0
flask-cors>=3.0.10
websockets>=10.0

# Machine learning (for AI models)
scikit-learn>=0.24.0
joblib>=1.0.0

# Type hints and validation
pydantic>=1.8.0
typing-extensions>=3.10.0

# Development tools
black>=21.5b0
flake8>=3.9.0
mypy>=0.910
isort>=5.9.0